Move Main pages to Views
